Week4/requirements.txt
polars
pandas
matplotlib
mplfinance
seaborn
scikit-learn
pytest
pytest-cov
mysql-connector-python
beautifulsoup4
requests
yfinance
streamlit
flake8
black
openai